+++ backend/life-insurance/src/main/java/bbd/miniconomy/lifeinsurance/repositories/TransactionRepository.java
@@ -3,5 +3,5 @@ package bbd.miniconomy.lifeinsurance.repositories;
 import bbd.miniconomy.lifeinsurance.models.entities.Transaction;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TransactionRepository extends JpaRepository<Transaction, Long> {
+public interface TransactionRepository extends JpaRepository<Transaction, Integer> {
 }
